Updated help, improved overall structure. Changed CLI usage.

+++ citationer.js
@@ -2,8 +2,8 @@ var argv = require("optimist").argv,
     fs = require("fs"),
     Citation = require("citation.js");
 
-var style = "MLA",
-    referenceFile;
+var style = "MLA";
+var referenceFile;
 
 if (argv.j || argv.json)
   style = "JSON";
@@ -14,6 +14,10 @@ if (fs.existsSync("citations.json"))
 if (argv.f || argv.file)
   referenceFile = argv.f || argv.file;
 
+/*
+  citationer init - create a citations.json
+  if none has already been created.
+*/
 if (process.argv[2] === "init") {
   if (!fs.existsSync("citations.json")) {
     try {
@@ -25,13 +29,18 @@ if (process.argv[2] === "init") {
     console.log("Created citations.json");
   }
   else {
-    console.error("citations.json already exists. Refusing to overwrite.");
+    console.error("citations.json already exists. Not overwriting.");
   }
 }
-else if (argv.s || argv.site) {
-  var site = argv.s || argv.site;
+/*
+  citationer [-s|--site] branches.io
+  Fetch the site.
+*/
+else if (argv.s || argv.site || argv._[0]) {
+  var site = argv.s || argv.site || argv._[0];
+  
   var URI;
-
+  
   if (!(/^https?\:\/\//.test(site)))
     URI = "http://" + site;
   else
@@ -40,20 +49,23 @@ else if (argv.s || argv.site) {
   var citation = new Citation(URI);
   citation.getReference(function(err, reference) {
     if (err) throw err;
+    
+    /* Always print to console */
+    
     if (style === "MLA")
       console.log(Citation.convertToMla(reference));
     else if (style === "JSON")
       console.log(JSON.stringify(reference, null, 2));
+    
+    /* And if a citations.json exits, write to it */
+    
     if (referenceFile) {
       try {
+        /* read reference file, and push new reference into it, write */
+        
         var references = JSON.parse(fs.readFileSync(referenceFile));
         references.push(reference)
-        try {
-          fs.writeFileSync(referenceFile, JSON.stringify(references, null, 2));
-        }
-        catch (err) {
-          throw err;
-        }
+        fs.writeFileSync(referenceFile, JSON.stringify(references, null, 2));
       } 
       catch (err) {
         throw err;
@@ -61,38 +73,53 @@ else if (argv.s || argv.site) {
     }
   });
 }
+/*
+citationer -l|--list
+*/
 else if (argv.l || argv.list) {
-  var citations = JSON.parse(fs.readFileSync(referenceFile));
-  var citer = new Citation();
-  citations.forEach(function(citation) {
-    console.log(citer.convertToMla(citation));
-  });
+  if (fs.existsSync(referenceFile)) {
+    var citations = JSON.parse(fs.readFileSync(referenceFile));
+    var citer = new Citation();
+    citations.forEach(function(citation) {
+      if (style === "MLA")
+        console.log(Citation.convertToMla(citation));
+      else if (style === "JSON")
+        console.log(citation);
+    });
+  }
 }
+/*
+citationer -e|--export [name]
+Export the citations as MLA.
+*/
 else if (argv.e || argv.export) {
   var exportfile = (argv.e || argv.exportfile);
-  console.log("export:", exportfile);
-  var citations = JSON.parse(fs.readFileSync(referenceFile));
-  var citer = new Citation();
-  var output = "";
-  citations.forEach(function(citation) {
-    output += citer.convertToMla(citation) + "\n";
-  });
-  try {
+
+  if (fs.existsSync(referenceFile)) {
+    var citations = JSON.parse(fs.readFileSync(referenceFile));
+    var output = "";
+    
+    citations.forEach(function(citation) {
+      output += Citation.convertToMla(citation) + "\n";
+    });
+    
+    try {
+      if (typeof(exportfile) === "string")
+        fs.writeFileSync(exportfile, output);
+      else
+        fs.writeFileSync(__dirname + "/citations.txt", output);
+    }
+    catch (err) {
+      throw err;
+    }
+    
     if (typeof(exportfile) === "string")
-      fs.writeFileSync(exportfile, output);
+      console.log("Exported as", exportfile);
     else
-      fs.writeFileSync(__dirname + "/citations.txt", output);
+      console.log("Exported as citations.txt");
   }
-  catch (err) {
-    throw err;
-  }
-  if (typeof(exportfile) === "string")
-    console.log("Exported as", exportfile);
-  else
-    console.log("Exported as citations.txt");
 }
-else if (process.argv[2] !== "init") {
+else if (argv.h || argv.help || true) {
   fs.createReadStream(__dirname + "/help.txt")
   .pipe(process.stdout);
 }
-
